MINOR: Remove Redundant BiValue Interface and Rename Abstract Class Accordingly
